+++ framework/model/src/main/java/is/codion/framework/model/EntityTableModel.java
@@ -239,12 +239,6 @@ public interface EntityTableModel<E extends EntityEditModel> extends FilteredMod
 	State conditionRequired();
 
 	/**
-	 * @return the state controlling whether this table model handles entity edit events, by replacing foreign key values
-	 * @see EntityEditEvents
-	 */
-	State editEvents();
-
-	/**
 	 * @return the Value controlling the action to perform when entities are inserted via the associated edit model
 	 * @see #ON_INSERT
 	 */